Block event fiber when waiting for X event
We can use IO::FileDescriptor#wait_readable (an undocumented but public
method) to yield a fiber until the file descriptor has something to read.

On POSIX systems (only), the X11 socket is exposed as a file descriptor.
That's an acceptable limitation for now---I don't care to support X11 on
Windows.

https://github.com/crystal-lang/crystal/pull/7366
https://github.com/crystal-lang/crystal/pull/8651
https://forum.crystal-lang.org/t/fibers-blocking-io-and-c-libraries/3116/13?u=elebow