Size and position all windows at startup
Except the taskbar, of course. We may need to exclude the taskbar some
other way when we implement the "uncategorized" and "all" categories,
because we don't ever want to focus the taskbar.

It's not so easy as adding a #reject to Minyaty::X.query_windows,
because that method is called by Taskbar's constructor (via
Categories#refresh) at which time we don't yet know what to reject.